// src/models/interfaces/theme.ts
export type Color =
  | "info"
  | "white"
  | "transparent"
  | "primary"
  | "danger"
  | "success"
  | "dark"
  | "warning"
  | "light"
  | "light-info"
  | "light-dark"
  | "light-primary"
  | "light-danger"
  | "light-success"
  | "light-warning";
export type ButtonVariant =
  | "outline"
  | "link"
  | "contained"
  | "light"
  | "transparent"
  | "outline-dashed";
export type ButtonSize = "xs" | "sm" | "lg" | "md";
export type ComponentStatusVariant =
  | "success"
  | "info"
  | "warning"
  | "error"
  | undefined;
export type IconSize = Size;

export type Size =
  | "1"
  | "2"
  | "3"
  | "4"
  | "5"
  | "6"
  | "7"
  | "8"
  | "9"
  | "10"
  | "1x"
  | "2x"
  | "3x"
  | "4x"
  | "5x"
  | "2qx"
  | "3qx"
  | "4qx"
  | "5qx"
  | "2hx"
  | "3hx"
  | "4hx"
  | "5hx"
  | "2tx"
  | "3tx"
  | "4tx"
  | "5tx";
export type ComponentBackgroundVariant =
  | "white"
  | "transparent"
  | "light"
  | "dark"
  | "warning"
  | "primary"
  | "info"
  | "success"
  | "danger"
  | "light-warning"
  | "light-primary"
  | "light-info"
  | "light-success"
  | "light-danger"
  | undefined;
export type ShadowVariant = "shadow-xs" | "shadow-sm" | "shadow-lg";

export type CheckboxVariant = "outline" | "solid";
export type AlignVariant = "right" | "center" | "left";
export type GalleryTypeVariant = "IMAGE" | "VIDEO";
export type DialogVariant =
  | "info"
  | "primary"
  | "danger"
  | "default"
  | "success"
  | "dark"
  | "warning";
export type Opacity = 0 | 5 | 10 | 15 | 20 | 25 | 50 | 75 | 100;